fix: update getUsers() to return Observable<User[]>and add country validation for users over 21

+++ user-management/src/app/pages/user-management/components/user-list/user-list.component.ts
@@ -4,9 +4,9 @@ import { MatSort } from '@angular/material/sort';
 import { MatPaginator } from '@angular/material/paginator';
 import { MatTableDataSource } from '@angular/material/table';
 import { UserManagementService } from '../../services/user-management.service';
-import { Subscription } from 'rxjs';
 import { SnackbarService } from '../../../../core/components/snackbar/services/snackbar.service';
 import { SnackbarType } from '../../../../core/components/snackbar/models/snackbar-type';
+import { Subscription } from 'rxjs';
 
 @Component({
   selector: 'app-user-list',
@@ -26,6 +26,7 @@ export class UserListComponent {
     'actions'
   ];
   users!: MatTableDataSource<User>;
+  subscription$: Subscription[] = [];
 
   userService = inject(UserManagementService);
   snackbar = inject(SnackbarService);
@@ -35,10 +36,15 @@ export class UserListComponent {
   }
 
   getUsers() {
-    this.users = new MatTableDataSource(this.userService.getUsers());
-    this.users.paginator = this.paginator;
-    this.users.sort = this.sort;
-    this.paginator._intl.itemsPerPageLabel = 'Ejemplos por pagina: ';
+    this.subscription$ =[
+      ...this.subscription$,
+      this.userService.getUsers().subscribe(result => {
+        this.users = new MatTableDataSource(result);
+        this.users.paginator = this.paginator;
+        this.users.sort = this.sort;
+        this.paginator._intl.itemsPerPageLabel = 'Ejemplos por pagina: ';
+      })
+    ]    
   }
 
   applyFilter(event: Event) {
@@ -51,12 +57,17 @@ export class UserListComponent {
   }
 
   deleteUsers(userId: string){
-    if(this.userService.deleteUser(userId)){
-      this.snackbar.openCustomSnackbar("User deleted", SnackbarType.warning);
-      this.getUsers();
-      return;
-    }
+    if(userId){
+      if(this.userService.deleteUser(userId)){
+        this.snackbar.openCustomSnackbar("User deleted", SnackbarType.warning);
+        return;
+      }
+    }    
     this.snackbar.openCustomSnackbar("User wasn't deleted", SnackbarType.error);
+  }
+
+  ngOnDestroy() {
+    this.subscription$.forEach(sub => sub.unsubscribe());
   }
 
 }
